Remove Wyre
Replace some wyre plugin check with dummy plugin check to retain code logic for future use.
