+++ src/components/data/row/PaymentMethodRow.tsx
@@ -5,7 +5,7 @@ import { cacheStyles } from 'react-native-patina'
 import buyPluginJsonRaw from '../../../constants/plugins/buyPluginList.json'
 import { guiPlugins } from '../../../constants/plugins/GuiPlugins'
 import sellPluginJsonRaw from '../../../constants/plugins/sellPluginList.json'
-import { PaymentMethod } from '../../../controllers/action-queue/WyreClient'
+import { PaymentMethod } from '../../../controllers/action-queue/PaymentMethod'
 import { lstrings } from '../../../locales/strings'
 import { asGuiPluginJson } from '../../../types/GuiPluginTypes'
 import { getPartnerIconUri } from '../../../util/CdnUris'
