+++ src/components/scenes/Loans/LoanManageScene.tsx
@@ -9,10 +9,10 @@ import { sprintf } from 'sprintf-js'
 import { AAVE_SUPPORT_ARTICLE_URL_1S } from '../../../constants/aaveConstants'
 import { guiPlugins } from '../../../constants/plugins/GuiPlugins'
 import { makeActionProgram } from '../../../controllers/action-queue/ActionProgram'
+import { PaymentMethodsMap } from '../../../controllers/action-queue/PaymentMethod'
 import { dryrunActionProgram } from '../../../controllers/action-queue/runtime/dryrunActionProgram'
 import { ActionOp, ActionProgram } from '../../../controllers/action-queue/types'
 import { makeInitialProgramState } from '../../../controllers/action-queue/util/makeInitialProgramState'
-import { PaymentMethodsMap } from '../../../controllers/action-queue/WyreClient'
 import { runLoanActionProgram } from '../../../controllers/loan-manager/redux/actions'
 import { LoanProgramType } from '../../../controllers/loan-manager/store'
 import { LoanAccount } from '../../../controllers/loan-manager/types'
