+++ src/controllers/action-queue/PaymentMethod.ts
@@ -1,0 +1,13 @@
+export interface PaymentMethod {
+  id: string
+  name: string
+  defaultCurrency: string
+  status: string
+  supportsDeposit: boolean
+  supportsPayment: boolean
+  blockchains: { [key: string]: string }
+}
+
+export interface PaymentMethodsMap {
+  [key: string]: PaymentMethod
+}
