+++ src/types/routerTypes.tsx
@@ -22,7 +22,7 @@ import { WcConnectParams } from '../components/scenes/WcConnectScene'
 import { WcDisconnectParams } from '../components/scenes/WcDisconnectScene'
 import { WebViewSceneParams } from '../components/scenes/WebViewScene'
 import { ExchangedFlipInputAmounts } from '../components/themed/ExchangedFlipInput'
-import { PaymentMethod } from '../controllers/action-queue/WyreClient'
+import { PaymentMethod } from '../controllers/action-queue/PaymentMethod'
 import { BorrowEngine, BorrowPlugin } from '../plugins/borrow-plugins/types'
 import { FiatPluginAddressFormParams, FiatPluginSepaFormParams, FiatPluginSepaTransferParams } from '../plugins/gui/fiatPluginTypes'
 import { FiatPluginEnterAmountParams } from '../plugins/gui/scenes/FiatPluginEnterAmountScene'
